requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
huggingface_hub
pymupdf  <-- NEW: Handles PDF-to-Image conversion